Basic android auto implementation
