+++ feature/main/src/main/java/ru/stersh/youamp/feature/main/Di.kt
@@ -1,6 +1,6 @@
 package ru.stersh.youamp.feature.main
 
-import org.koin.androidx.viewmodel.dsl.viewModel
+import org.koin.core.module.dsl.viewModel
 import org.koin.dsl.module
 import ru.stersh.youamp.feature.main.data.ServerInfoRepositoryImpl
 import ru.stersh.youamp.feature.main.domain.ServerInfoRepository
